interim
moved more out of conn_log
Add internal header to log.py
