+++ src/conn/conn_log.c
@@ -358,151 +358,6 @@ __wti_logmgr_reconfig(WT_SESSION_IMPL *session, const char **cfg)
 }
 
 /*
- * __log_remove_once_int --
- *     Helper for __log_remove_once. Intended to be called while holding the hot backup read lock.
- */
-static int
-__log_remove_once_int(
-  WT_SESSION_IMPL *session, char **logfiles, u_int logcount, uint32_t min_lognum)
-{
-    uint32_t lognum;
-    u_int i;
-
-    for (i = 0; i < logcount; i++) {
-        WT_RET(__wt_log_extract_lognum(session, logfiles[i], &lognum));
-        if (lognum < min_lognum)
-            WT_RET(__wt_log_remove(session, WT_LOG_FILENAME, lognum));
-    }
-
-    return (0);
-}
-
-/*
- * __compute_min_lognum --
- *     Determine the number of the earliest log file we must keep.
- */
-static uint32_t
-__compute_min_lognum(WT_SESSION_IMPL *session, WT_LOG *log, uint32_t backup_file)
-{
-#ifdef HAVE_DIAGNOSTIC
-    struct timespec ts;
-#endif
-    WT_CONNECTION_IMPL *conn;
-    uint32_t min_lognum;
-
-    conn = S2C(session);
-
-    /*
-     * If we're coming from a backup cursor we want the smaller of the last full log file copied in
-     * backup or the checkpoint LSN. Otherwise we want the minimum of the last log file written to
-     * disk and the checkpoint LSN.
-     */
-    min_lognum = backup_file == 0 ? WT_MIN(log->ckpt_lsn.l.file, log->sync_lsn.l.file) :
-                                    WT_MIN(log->ckpt_lsn.l.file, backup_file);
-
-    __wt_readlock(session, &conn->debug_log_retention_lock);
-
-    /* Adjust the number of log files to retain based on debugging options. */
-
-    if (FLD_ISSET(conn->debug_flags, WT_CONN_DEBUG_CKPT_RETAIN) && conn->debug_ckpt_cnt != 0)
-        min_lognum = WT_MIN(conn->debug_ckpt[conn->debug_ckpt_cnt - 1].l.file, min_lognum);
-
-    if (conn->debug_log_cnt != 0) {
-        /*
-         * If we're performing checkpoints, apply the retain value as a minimum, increasing the
-         * number the log files we keep. If not performing checkpoints, it's an absolute number of
-         * log files to keep. This means we can potentially remove log files required for recovery
-         * if the number of log files exceeds the configured value and the system has yet to be
-         * checkpointed.
-         *
-         * Check for N+1, that is, we retain N full log files, and one partial.
-         */
-        if ((conn->debug_log_cnt + 1) >= log->fileid)
-            min_lognum = 0;
-        else if (WT_IS_INIT_LSN(&log->ckpt_lsn))
-            min_lognum = log->fileid - (conn->debug_log_cnt + 1);
-        else
-            min_lognum = WT_MIN(log->fileid - (conn->debug_log_cnt + 1), min_lognum);
-    }
-
-    __wt_readunlock(session, &conn->debug_log_retention_lock);
-#ifdef HAVE_DIAGNOSTIC
-    __wt_epoch(session, &ts);
-    if (min_lognum > WT_INIT_LSN_FILE && min_lognum != log->min_fileid) {
-        WT_IGNORE_RET(__wt_log_printf(session,
-          "COMPUTE_MIN_LOG: [%" PRIuMAX ":%" PRIuMAX "] min_lognum %" PRIu32 " ckpt file %" PRIu32
-          " sync file %" PRIu32 " backup_file %" PRIu32 " debug_log count%" PRIu32
-          " old min %" PRIu32,
-          (uintmax_t)ts.tv_sec, (uintmax_t)ts.tv_nsec / WT_THOUSAND, min_lognum,
-          log->ckpt_lsn.l.file, log->sync_lsn.l.file, backup_file, conn->debug_log_cnt,
-          log->min_fileid));
-        log->min_fileid = min_lognum;
-    }
-#endif
-    return (min_lognum);
-}
-
-/*
- * __log_remove_once --
- *     Perform one iteration of log removal. Must be called with the log removal lock held.
- */
-static int
-__log_remove_once(WT_SESSION_IMPL *session, uint32_t backup_file)
-{
-    WT_CONNECTION_IMPL *conn;
-    WT_DECL_RET;
-    WT_LOG *log;
-    uint32_t min_lognum;
-    u_int logcount;
-    char **logfiles;
-
-    conn = S2C(session);
-    log = conn->log;
-    logcount = 0;
-    logfiles = NULL;
-
-    min_lognum = __compute_min_lognum(session, log, backup_file);
-    if (min_lognum <= WT_INIT_LSN_FILE)
-        /*
-         * If the minimum log number is zero or one that means we want to keep all logs and there is
-         * nothing to do here. In typical operation, the minimum returned will be one as that is the
-         * initial LSN file value. When using debug mode, the minimum returned may be zero.
-         */
-        return (0);
-
-    __wt_verbose(session, WT_VERB_LOG, "log_remove: remove to log number %" PRIu32, min_lognum);
-
-    /*
-     * Main remove code. Get the list of all log files and remove any earlier than the minimum log
-     * number.
-     */
-    WT_ERR(__wt_fs_directory_list(session, conn->log_path, WT_LOG_FILENAME, &logfiles, &logcount));
-
-    /*
-     * If backup_file is non-zero we know we're coming from an incremental backup cursor. In that
-     * case just perform the remove operation without the lock.
-     */
-    if (backup_file != 0)
-        ret = __log_remove_once_int(session, logfiles, logcount, min_lognum);
-    else
-        WT_WITH_HOTBACKUP_READ_LOCK(
-          session, ret = __log_remove_once_int(session, logfiles, logcount, min_lognum), NULL);
-    WT_ERR(ret);
-
-    /*
-     * Indicate what is our new earliest LSN. It is the start of the log file containing the last
-     * checkpoint.
-     */
-    WT_SET_LSN(&log->first_lsn, min_lognum, 0);
-
-    if (0)
-err:
-        __wt_err(session, ret, "log removal server error");
-    WT_TRET(__wt_fs_directory_list_free(session, &logfiles, logcount));
-    return (ret);
-}
-
-/*
  * __log_prealloc_once --
  *     Perform one iteration of log pre-allocation.
  */
@@ -564,42 +419,6 @@ __log_prealloc_once(WT_SESSION_IMPL *session)
 err:
         __wt_err(session, ret, "log pre-alloc server error");
     WT_TRET(__wt_fs_directory_list_free(session, &recfiles, reccount));
-    return (ret);
-}
-
-/*
- * __wt_log_truncate_files --
- *     Truncate log files via remove once. Requires that the server is not currently running.
- */
-int
-__wt_log_truncate_files(WT_SESSION_IMPL *session, WT_CURSOR *cursor, bool force)
-{
-    WT_CONNECTION_IMPL *conn;
-    WT_DECL_RET;
-    WT_LOG *log;
-    uint32_t backup_file;
-
-    conn = S2C(session);
-    if (!FLD_ISSET(conn->log_flags, WT_CONN_LOG_ENABLED))
-        return (0);
-    if (!force && FLD_ISSET(conn->server_flags, WT_CONN_SERVER_LOG) &&
-      FLD_ISSET(conn->log_flags, WT_CONN_LOG_REMOVE))
-        WT_RET_MSG(session, EINVAL, "Attempt to remove manually while a server is running");
-
-    log = conn->log;
-
-    backup_file = 0;
-    if (cursor != NULL) {
-        WT_ASSERT(session, force == false);
-        backup_file = WT_CURSOR_BACKUP_ID(cursor);
-    }
-    WT_ASSERT(session, backup_file <= log->alloc_lsn.l.file);
-    __wt_verbose(
-      session, WT_VERB_LOG, "log_truncate_files: remove once up to %" PRIu32, backup_file);
-
-    __wt_writelock(session, &log->log_remove_lock);
-    ret = __log_remove_once(session, backup_file);
-    __wt_writeunlock(session, &log->log_remove_lock);
     return (ret);
 }
 
@@ -688,145 +507,6 @@ err:
     }
     __wt_spin_unlock_if_owned(session, &log->log_sync_lock);
     return (WT_THREAD_RET_VALUE);
-}
-
-/*
- * Simple structure for sorting written slots.
- */
-typedef struct {
-    WT_LSN lsn;
-    uint32_t slot_index;
-} WT_LOG_WRLSN_ENTRY;
-
-/*
- * WT_WRLSN_ENTRY_CMP_LT --
- *	Return comparison of a written slot pair by LSN.
- */
-#define WT_WRLSN_ENTRY_CMP_LT(entry1, entry2)        \
-    ((entry1).lsn.l.file < (entry2).lsn.l.file ||    \
-      ((entry1).lsn.l.file == (entry2).lsn.l.file && \
-        __wt_lsn_offset(&(entry1).lsn) < __wt_lsn_offset(&(entry2).lsn)))
-
-/*
- * __wt_log_wrlsn --
- *     Process written log slots and attempt to coalesce them if the LSNs are contiguous. The
- *     purpose of this function is to advance the write_lsn in LSN order after the buffer is written
- *     to the log file.
- */
-void
-__wt_log_wrlsn(WT_SESSION_IMPL *session, int *yield)
-{
-    WT_CONNECTION_IMPL *conn;
-    WT_LOG *log;
-    WT_LOGSLOT *coalescing, *slot;
-    WT_LOG_WRLSN_ENTRY written[WT_SLOT_POOL];
-    WT_LSN save_lsn;
-    size_t written_i;
-    uint32_t i, save_i, slot_last_offset;
-
-    conn = S2C(session);
-    log = conn->log;
-    __wt_spin_lock(session, &log->log_writelsn_lock);
-restart:
-    coalescing = NULL;
-    WT_INIT_LSN(&save_lsn);
-    written_i = 0;
-    i = 0;
-
-    /*
-     * Walk the array once saving any slots that are in the WT_LOG_SLOT_WRITTEN state.
-     */
-    while (i < WT_SLOT_POOL) {
-        save_i = i;
-        slot = &log->slot_pool[i++];
-        if (__wt_atomic_loadiv64(&slot->slot_state) != WT_LOG_SLOT_WRITTEN)
-            continue;
-        written[written_i].slot_index = save_i;
-        WT_ASSIGN_LSN(&written[written_i++].lsn, &slot->slot_release_lsn);
-    }
-    /*
-     * If we found any written slots process them. We sort them based on the release LSN, and then
-     * look for them in order.
-     */
-    if (written_i > 0) {
-        if (yield != NULL)
-            *yield = 0;
-        WT_INSERTION_SORT(written, written_i, WT_LOG_WRLSN_ENTRY, WT_WRLSN_ENTRY_CMP_LT);
-        /*
-         * We know the written array is sorted by LSN. Go through them either advancing write_lsn or
-         * coalesce contiguous ranges of written slots.
-         */
-        for (i = 0; i < written_i; i++) {
-            slot = &log->slot_pool[written[i].slot_index];
-            /*
-             * The log server thread pushes out slots periodically. Sometimes they are empty slots.
-             * If we find an empty slot, where empty means the start and end LSN are the same, free
-             * it and continue.
-             */
-            if (__wt_log_cmp(&slot->slot_start_lsn, &slot->slot_release_lsn) == 0 &&
-              __wt_log_cmp(&slot->slot_start_lsn, &slot->slot_end_lsn) == 0) {
-                __wt_log_slot_free(session, slot);
-                continue;
-            }
-            if (coalescing != NULL) {
-                /*
-                 * If the write_lsn changed, we may be able to process slots. Try again.
-                 */
-                if (__wt_log_cmp(&log->write_lsn, &save_lsn) != 0)
-                    goto restart;
-                if (__wt_log_cmp(&coalescing->slot_end_lsn, &written[i].lsn) != 0) {
-                    coalescing = slot;
-                    continue;
-                }
-                /*
-                 * If we get here we have a slot to coalesce and free.
-                 */
-                __wt_atomic_storeiv64(
-                  &coalescing->slot_last_offset, __wt_atomic_loadi64(&slot->slot_last_offset));
-                WT_ASSIGN_LSN(&coalescing->slot_end_lsn, &slot->slot_end_lsn);
-                WT_STAT_CONN_INCR(session, log_slot_coalesced);
-                /*
-                 * Copy the flag for later closing.
-                 */
-                if (F_ISSET_ATOMIC_16(slot, WT_SLOT_CLOSEFH))
-                    F_SET_ATOMIC_16(coalescing, WT_SLOT_CLOSEFH);
-            } else {
-                /*
-                 * If this written slot is not the next LSN, try to start coalescing with later
-                 * slots. A synchronous write may update write_lsn so save the last one we saw to
-                 * check when coalescing slots.
-                 */
-                WT_ASSIGN_LSN(&save_lsn, &log->write_lsn);
-                if (__wt_log_cmp(&log->write_lsn, &written[i].lsn) != 0) {
-                    coalescing = slot;
-                    continue;
-                }
-                /*
-                 * If we get here we have a slot to process. Advance the LSN and process the slot.
-                 */
-                WT_ASSERT(session, __wt_log_cmp(&written[i].lsn, &slot->slot_release_lsn) == 0);
-                /*
-                 * We need to maintain the starting offset of a log record so that the checkpoint
-                 * LSN refers to the beginning of a real record. The last offset in a slot is kept
-                 * so that the checkpoint LSN is close to the end of the record.
-                 */
-                slot_last_offset = (uint32_t)__wt_atomic_loadi64(&slot->slot_last_offset);
-                if (__wt_lsn_offset(&slot->slot_start_lsn) != slot_last_offset)
-                    __wt_atomic_store32(&slot->slot_start_lsn.l.offset, slot_last_offset);
-                WT_ASSIGN_LSN(&log->write_start_lsn, &slot->slot_start_lsn);
-                WT_ASSIGN_LSN(&log->write_lsn, &slot->slot_end_lsn);
-                __wt_cond_signal(session, log->log_write_cond);
-                WT_STAT_CONN_INCR(session, log_write_lsn);
-                /*
-                 * Signal the close thread if needed.
-                 */
-                if (F_ISSET_ATOMIC_16(slot, WT_SLOT_CLOSEFH))
-                    __wt_cond_signal(session, conn->log_file_cond);
-            }
-            __wt_log_slot_free(session, slot);
-        }
-    }
-    __wt_spin_unlock(session, &log->log_writelsn_lock);
 }
 
 /*
@@ -956,7 +636,7 @@ __log_server(void *arg)
              */
             if (FLD_ISSET(conn->log_flags, WT_CONN_LOG_REMOVE)) {
                 if (__wt_try_writelock(session, &log->log_remove_lock) == 0) {
-                    ret = __log_remove_once(session, 0);
+                    ret = __wt_log_remove_once(session, 0);
                     __wt_writeunlock(session, &log->log_remove_lock);
                     WT_ERR(ret);
                 } else
